baseline/api-service/requirements.txt
Flask==2.0.1
psycopg2-binary>=2.9.9
requests==2.25.1
PyJWT==1.7.1
Werkzeug==2.0.1
click==7.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
